refactor(mode): split default-mode UI into ui/default_mode, session keeps lifecycle
