+++ src/session/default/spawn.rs
@@ -10,14 +10,14 @@ use std::path::PathBuf;
 ///
 /// `None` means fall back to `$SHELL` at spawn time.
 #[derive(Resource)]
-pub(super) struct OzmaTerminalConfig {
+pub(crate) struct OzmaTerminalConfig {
     /// Optional shell path. When set, overrides `$SHELL` and `/bin/sh`.
     pub shell: Option<String>,
 }
 
 /// Options for spawning a standalone Ozma terminal.
 #[derive(Default)]
-pub(super) struct OzmaSpawnOptions {
+pub(crate) struct OzmaSpawnOptions {
     /// Shell override; `None` falls back to `$SHELL` then `/bin/sh`.
     pub shell: Option<String>,
     /// Working directory for the PTY; `None` inherits the process cwd.
@@ -31,7 +31,7 @@ pub(super) struct OzmaSpawnOptions {
 /// GPU render bundle is injected by `crate::surface`'s add-observer on
 /// insertion.
 #[derive(Bundle)]
-pub(super) struct OzmaTerminalBundle {
+pub(crate) struct OzmaTerminalBundle {
     terminal: TerminalBundle,
     marker: OzmaTerminal,
     node: Node,
